requirements.txt
pandas
# random
numpy
itertools
plotly.express
